requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
matplotlib
jinja2
numpy
